RIB-51: Removed redundant method, add condition creator cant delete creator

+++ src/main/java/it/rate/webapp/services/ManageInterestService.java
@@ -48,8 +48,12 @@ public class ManageInterestService {
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Role not found");
     }
     Role role = optRole.get();
+    if (role.getRole().equals(Role.RoleType.CREATOR)) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Cannot remove creator role");
+    }
     RoleId roleId = new RoleId(role.getId().getUserId(), role.getId().getInterestId());
     roleService.deleteByRoleId(roleId);
+    // todo : ADD TESTS FOR CREATOR NOT ABLE TO REMOVE CREATOR
     // todo : CHANGES WONT TAKE EFFECT UNTIL USER RELOGS // NEED TO MANIPULATE SESSION
   }
